feat: refactoring of classes, dto separation and added relatoriorestaurante

+++ src/main/java/com/deliverytech/delivery/dto/request/PedidoRequestDTO.java
@@ -4,10 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
+import java.math.BigDecimal; 
 
 /**
  * DTO para o cadastro de novos pedidos.
- * Contém apenas os campos que o cliente deve informar.
  */
 @Data
 public class PedidoRequestDTO {
@@ -19,12 +19,11 @@ public class PedidoRequestDTO {
     private Long restauranteId;
 
     @NotBlank(message = "A lista de itens é obrigatória")
-    private String itens; // Mantido como String, conforme a entidade
+    private String itens;
 
     @NotNull(message = "O valor total é obrigatório")
     @Positive(message = "O valor total deve ser maior que zero")
-    private Double valorTotal;
+    private BigDecimal valorTotal; 
 
-    // Observações são opcionais
     private String observacoes;
 }
