+++ src/main/java/com/deliverytech/delivery/dto/request/ProdutoRequestDTO.java
@@ -4,10 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
+import java.math.BigDecimal;
 
 /**
  * DTO para o cadastro e atualização de Produtos.
- * Contém os campos que o cliente deve informar, com as devidas validações.
  */
 @Data
 public class ProdutoRequestDTO {
@@ -15,12 +15,11 @@ public class ProdutoRequestDTO {
     @NotBlank(message = "O nome é obrigatório")
     private String nome;
 
-    // Descrição é opcional
     private String descricao;
 
     @NotNull(message = "O preço é obrigatório")
     @Positive(message = "O preço deve ser maior que zero")
-    private Double preco;
+    private BigDecimal preco;
 
     // Categoria é opcional
     private String categoria;
@@ -28,7 +27,4 @@ public class ProdutoRequestDTO {
     @NotNull(message = "O ID do restaurante é obrigatório")
     private Long restauranteId;
 
-    // O campo 'disponivel' é gerenciado pelo service:
-    // - Definido como 'true' no cadastro.
-    // - Alterado para 'false' pelo endpoint DELETE (tornarIndisponivel).
 }
