+++ src/main/java/com/deliverytech/delivery/entity/Pedido.java
@@ -1,12 +1,12 @@
 package com.deliverytech.delivery.entity;
 
-import com.deliverytech.delivery.enums.StatusPedido; // Import adicionado
+import com.deliverytech.delivery.enums.StatusPedido; 
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;     // Import adicionado
-import jakarta.persistence.Enumerated; // Import adicionado
+import jakarta.persistence.EnumType;     
+import jakarta.persistence.Enumerated; 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -33,12 +33,11 @@ public class Pedido {
     @Column(name = "data_pedido")
     private LocalDateTime dataPedido;
 
-    // Alterado de String para StatusPedido
     @Enumerated(EnumType.STRING)
     private StatusPedido status;
 
     @Column(name = "valor_total")
-    private Double valorTotal;
+    private BigDecimal valorTotal;
 
     private String observacoes;
