+++ src/main/java/com/deliverytech/delivery/repository/ProdutoRepository.java
@@ -4,7 +4,7 @@ import com.deliverytech.delivery.entity.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-// import java.math.BigDecimal; // Removido
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -12,13 +12,11 @@ public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 
     /**
      * Buscar produtos disponíveis
-     * (Equivalente ao findByAtivoTrue de Cliente)
      */
     List<Produto> findByDisponivelTrue();
 
     /**
      * Buscar produtos por nome (contendo)
-     * (Padrão de ClienteRepository)
      */
     List<Produto> findByNomeContainingIgnoreCase(String nome);
 
@@ -38,8 +36,7 @@ public interface ProdutoRepository extends JpaRepository<Produto, Long> {
     List<Produto> findByCategoriaContainingIgnoreCase(String categoria);
 
     /**
-     * CORRIGIDO: Buscar produtos com preço menor ou igual ao valor.
-     * (Trocado BigDecimal por Double para bater com a Entidade)
-     */
-    List<Produto> findByPrecoLessThanEqual(Double preco); // Corrigido
+     * Buscar produtos com preço menor ou igual ao valor.
+    */
+    List<Produto> findByPrecoLessThanEqual(BigDecimal preco); 
 }
